+++ Lib9c.Renderers/Lib9c.Renderers.csproj
@@ -8,7 +8,5 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Lib9c\Lib9c.csproj" />
-    <ProjectReference Include="..\Lib9c.Renderers\Lib9c.Renderers.csproj" />
   </ItemGroup>
-
 </Project>
