+++ Lib9c.sln
@@ -45,6 +45,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lib9c.Miner.Tests", ".Lib9c
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lib9c.Policy", "Lib9c.Policy\Lib9c.Policy.csproj", "{7DF9FB47-1B56-4F4E-BA98-4F41C4B0CBF8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lib9c.Renderers", "Lib9c.Renderers\Lib9c.Renderers.csproj", "{321113A4-7CAA-462C-941E-366EBD4668EB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -160,6 +162,12 @@ Global
 		{7DF9FB47-1B56-4F4E-BA98-4F41C4B0CBF8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7DF9FB47-1B56-4F4E-BA98-4F41C4B0CBF8}.DevEx|Any CPU.ActiveCfg = Debug|Any CPU
 		{7DF9FB47-1B56-4F4E-BA98-4F41C4B0CBF8}.DevEx|Any CPU.Build.0 = Debug|Any CPU
+		{321113A4-7CAA-462C-941E-366EBD4668EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{321113A4-7CAA-462C-941E-366EBD4668EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{321113A4-7CAA-462C-941E-366EBD4668EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{321113A4-7CAA-462C-941E-366EBD4668EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{321113A4-7CAA-462C-941E-366EBD4668EB}.DevEx|Any CPU.ActiveCfg = Debug|Any CPU
+		{321113A4-7CAA-462C-941E-366EBD4668EB}.DevEx|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
